sqlite_to_postgres/pyproject.toml
[tool.poetry]
name = "sqlite-to-postgres"
version = "0.0.1"
description = "Script for migrating SQLite db data to Postgres db"
authors = ["shitcoding <shitcoder@pm.me>"]
readme = "README.md"
packages = [{include = "sqlite_to_postgres"}]

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "2.9.7"
python-dotenv = "^1.0.0"
loguru = "^0.7.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
